Creates the initial project version.
Creates the core project files and folders, and includes
the first versions of all README documentation for this
repository.

Adds challege_setup/ with Python scripts and a requirements.txt
file in the root directory, specifically for Python. PowerShell
users can ignore this folder and file unless they are curious.

Paves way for GitHub actions workflows to be created.